turn on multiple projects support

+++ server-side/localrun.sh
@@ -1,21 +1,19 @@
 set -e
-# USAGE: specify link to remote repo as first arg. For example: ./localrun.sh git@github.com:ilya-t/wiki-app.git
 PORT=8181
-REPO=$1
+host_volumes=$1
 
-if [ "$REPO" == "" ]; then
+if [ "$host_volumes" == "" ]; then
     REPO="./tests/test_repo"
-    echo "WARNING! NO REPO SPECIFIED! USING TEST REPO AT: $REPO"
-    cd tests
-    ./init_repo.sh
-    cd ..
+    echo "Specify path to volumes. Expecting stucture:"
+    echo "  /some/path/config/config.json"
+    echo "  /some/path/repo-store"
+    exit 1
 fi
 
-target_ssh_keys=~/.ssh
+host_ssh_keys=~/.ssh
 
-# if [[ -d "$REPO" ]]; then
-#     # TODO: mount bare local repo for cloning?
-# fi
+mkdir -p $host_volumes/config
+mkdir -p $host_volumes/repo-store
 
 docker build ./app --tag wiki_backend
 docker container rm --force wiki_backend_local
@@ -24,7 +22,9 @@ docker run \
     --detach \
     --publish $PORT:80 \
     --env APP_REPO_LINK=$REPO \
-    --volume $target_ssh_keys:/root/.ssh \
+    --volume $host_ssh_keys:/root/.ssh \
+    --volume $host_volumes/config:/app/config \
+    --volume $host_volumes/repo-store:/app/repo-store \
     --name wiki_backend_local \
     wiki_backend:latest
